requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
python-dateutil>=2.8.2
backports.zoneinfo>=0.2.1